+++ lets/modules/encode/powershell/rc4.py
@@ -2,11 +2,13 @@ from lets.module import Module
 from lets.extensions.docker import DockerExtension
 
 # Imports required to execute this module
-import base64, string
+import os, gzip, base64, string
+from Crypto.Cipher import ARC4
+from Crypto.Hash import SHA256
 
-class Base64(DockerExtension, Module):
+class Rc4(DockerExtension, Module):
     """
-    Base64 encode a powershell script and prepend a decode stub.
+    Compress, Encrypt and Base64 encode a powershell script and prepend a decode/decrypt/decompress stub.
     """
 
     # A list of docker images required by the module.
@@ -21,6 +23,11 @@ class Base64(DockerExtension, Module):
         :return: ArgumentParser object
         """
         parser = super().usage()
+
+        # Set encryption parameters
+        parser.add_argument("-k", "--key",
+            help="key used for RC4 encryption",
+            type=str)
 
         return parser
 
@@ -44,12 +51,29 @@ class Base64(DockerExtension, Module):
         except AssertionError as e:
             self.throw(e)
 
+        # Normalize key
+        if self.options.get("key"):
+            key = SHA256.new(self.options.get("key").encode()).digest()
+        else:
+            key = os.urandom(32)
+
+        # Compress command
+        compressed = gzip.compress(data, compresslevel=9)
+
+        # Encrypt command
+        encrypted = ARC4.new(key).encrypt(compressed)
+
         # Encode command
-        encoded = base64.b64encode(data)
+        encoded = base64.b64encode(encrypted)
         self.options["encoded"] = encoded.decode()
+
+        # Encode key
+        encoded_key = base64.b64encode(key)
+        self.options["encoded_key"] = encoded_key.decode()
         
         # Place encoded command in harness
-        cmd = "IEX ([System.Text.Encoding]::UTF8.GetString([System.Convert]::FromBase64String('%(encoded)s')))" % self.options
+        self.options["encrypted"] = "$(& $({$D,$K=$Args;$S=0..255;0..255|%%{$J=($J+$S[$_]+$K[$_%%$K.Length])%%256;$S[$_],$S[$J]=$S[$J],$S[$_]};$D|%%{$I=($I+1)%%256;$H=($H+$S[$I])%%256;$S[$I],$S[$H]=$S[$H],$S[$I];$_-bxor$S[($S[$I]+$S[$H])%%256]}}) $([System.Convert]::FromBase64String('%(encoded)s')) $([System.Convert]::FromBase64String('%(encoded_key)s')))" % self.options
+        cmd = "IEX (New-Object System.IO.StreamReader($(New-Object System.IO.Compression.GzipStream($(New-Object System.IO.MemoryStream(,%(encrypted)s)),[System.IO.Compression.CompressionMode]::Decompress)),[System.Text.Encoding]::UTF8)).ReadToEnd()" % self.options
 
         # Convert harness to bytes and return
         yield cmd.encode()
@@ -58,11 +82,14 @@ class Base64(DockerExtension, Module):
         """
         Perform unit tests to verify this module's functionality.
         """
-        # Test encoding
+        # Test key encoding
         self.assertTrue(
-            b"AAECAwQFBgcICQoLDA0ODxAREhMUFRYXGBkaGxwdHh8gISIjJCUmJygpKissLS4vMDEyMzQ1Njc4OTo7PD0+P0BBQkNERUZHSElKS0xNTk9QUVJTVFVWV1hZWltcXV5fYGFiY2RlZmdoaWprbG1ub3BxcnN0dXZ3eHl6e3x9fn+AgYKDhIWGh4iJiouMjY6PkJGSk5SVlpeYmZqbnJ2en6ChoqOkpaanqKmqq6ytrq+wsbKztLW2t7i5uru8vb6/wMHCw8TFxsfIycrLzM3Oz9DR0tPU1dbX2Nna29zd3t/g4eLj5OXm5+jp6uvs7e7v8PHy8/T19vf4+fr7/P3+/w==" in 
-            b"".join(self.do(bytes(range(0, 256)))),
-            "All bytes produced innacurate results")
+            b"IqSAUVlMGUne7XBAhQwfD4dkU39Rkb5Wcy0WpUwdgVM=" in 
+            b"".join(self.do(bytes(range(0, 256)), {"key": "A"*32})),
+            "Key encoding produced innacurate results")
+
+        # Test content encryption (gzip header not consistent)
+        # (skip, unreliable output)
 
         # Test execution
         test = string.ascii_letters + string.digits